Fully rework angle_between, and make other fixes
Rework angle_between based on http://www.euclideanspace.com/maths/algebra/vectors/angleBetween/index.htm
Fix argument order in the calls to angle_between in find_next_vertex
Make find_next_start_edge vertex order consistent, so that the "clockwise" argument defines the direction of the polygon internal angles
Reorganize some of the tests, and make them account for the new implementation of angle_between
Add length method to Vector2 in case it's needed. I added this because one of the ways to calculate the angle between two connected line segments needs the length of the line

I hope I'm not just throwing rocks into the mud and seeing whether or not they stick... Compared to this, it seems like my jsdoom code worked on flukes. All the tests pass without failing or panicking, so that's good.