Use OpenGL vertex array objects properly
The vertex array object stores information about which buffers and attributes to use, so the only object which must be bound before the draw call is the vertex array object.

Make Renderable::setup return a Result, because it could potentially fail, although it should not in most cases.

Also, add a "Full path" check box to the texture browser window.